+++ samples/SvgRecipeDemo/MainWindow.axaml.cs
@@ -15,6 +15,7 @@ using Avalonia.Threading;
 using SkiaSharp;
 using Svg.CodeGen.Skia;
 using Svg.CodeGen.Skia.Expressions;
+using Svg.Expressions;
 
 namespace SvgRecipeDemo;