+++ tests/Svg.Skia.UnitTests/ExprCompilerTests.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using Svg.CodeGen.Skia.Expressions;
+using Svg.Expressions;
 using Xunit;
 
 namespace Svg.Skia.UnitTests;