+++ tests/Svg.Skia.UnitTests/SkiaCSharpSingleFileTests.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using ShimSkiaSharp;
 using Svg.CodeGen.Skia;
 using Svg.CodeGen.Skia.Expressions;
+using Svg.Expressions;
 using Svg.Model.Services;
 using Xunit;